# docs/openapi.json
{"openapi":"3.1.0","info":{"contact":{"name":"PTITMeet Team"},"description":"API documentation for the PTITMeet backend services.","title":"PTITMeet API","version":"v1"},"servers":[{"url":"http://localhost:8080/api","description":"PTITMeet backend server"}],"paths":{"/api/auth/forgot-password":{"post":{"operationId":"forgotPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}},"description":"OK"}},"tags":["auth-controller"]}},"/api/auth/google":{"post":{"operationId":"loginWithGoogle","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GoogleLoginRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}},"description":"OK"}},"tags":["auth-controller"]}},"/api/auth/login":{"post":{"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}},"description":"OK"}},"tags":["auth-controller"]}},"/api/auth/logout":{"post":{"operationId":"logout","responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}},"description":"OK"}},"tags":["auth-controller"]}},"/api/auth/refresh-token":{"post":{"operationId":"refreshToken","responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseAuthResponse"}}},"description":"OK"}},"tags":["auth-controller"]}},"/api/auth/register":{"post":{"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}},"description":"OK"}},"tags":["auth-controller"]}},"/api/auth/reset-password":{"post":{"operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}},"description":"OK"}},"tags":["auth-controller"]}},"/api/livekit/recordings/my":{"get":{"operationId":"getMyRecordings","responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListMeetingRecording"}}},"description":"OK"}},"tags":["recording-controller"]}},"/api/livekit/recordings/start":{"post":{"operationId":"startRecording","parameters":[{"in":"query","name":"meetingCode","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMeetingRecording"}}},"description":"OK"}},"tags":["recording-controller"]}},"/api/livekit/recordings/status":{"get":{"operationId":"getRecordingStatus","parameters":[{"in":"query","name":"egressId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["recording-controller"]}},"/api/livekit/recordings/stop":{"post":{"operationId":"stopRecording","parameters":[{"in":"query","name":"egressId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["recording-controller"]}},"/api/livekit/webhook":{"post":{"operationId":"handleWebhook","parameters":[{"in":"header","name":"Authorization","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"type":"string"}}},"description":"OK"}},"tags":["live-kit-webhook-controller"]}},"/api/meetings/history":{"get":{"operationId":"getHistory","parameters":[{"in":"query","name":"page","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"in":"query","name":"size","required":false,"schema":{"type":"integer","format":"int32","default":6}},{"in":"query","name":"role","required":false,"schema":{"type":"string","default":"ALL"}},{"in":"query","name":"status","required":false,"schema":{"type":"string","default":"ALL"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponsePageMeetingHistoryResponse"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/instant":{"post":{"operationId":"createInstant","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMeetingRequest"}}}},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMeeting"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/my-meetings":{"get":{"operationId":"getMyMeetings","responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListMeeting"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/schedule":{"post":{"operationId":"schedule","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMeetingRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMeeting"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/up-next":{"get":{"operationId":"getUpNext","responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMeetingHistoryResponse"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/{code}":{"delete":{"operationId":"cancel","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/{code}/approval":{"post":{"operationId":"approveParticipant","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApprovalRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/{code}/chat/history":{"get":{"operationId":"getChatHistory","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListChatMessage"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/{code}/end":{"post":{"operationId":"endForAll","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/{code}/feedback":{"post":{"operationId":"submitFeedback","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/{code}/info":{"get":{"operationId":"getInfo","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMeetingInfoResponse"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/{code}/join":{"post":{"operationId":"joinMeeting","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JoinMeetingRequest"}}}},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseJoinMeetingResponse"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/{code}/leave":{"post":{"operationId":"leave","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseVoid"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/{code}/settings":{"get":{"operationId":"getSettings","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}},"description":"OK"}},"tags":["meeting-controller"]},"put":{"operationId":"updateSettings","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{}}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMeeting"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/{code}/summary":{"get":{"operationId":"getSummary","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}},{"in":"query","name":"action","required":false,"schema":{"type":"string","default":"LEAVE"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseMeetingSummaryResponse"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/meetings/{code}/waiting-room":{"get":{"operationId":"getWaitingRoom","parameters":[{"in":"path","name":"code","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListParticipantResponse"}}},"description":"OK"}},"tags":["meeting-controller"]}},"/api/users/avatar":{"post":{"operationId":"uploadAvatar","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]}}}},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}},"description":"OK"}},"tags":["user-controller"]}},"/api/users/me":{"get":{"operationId":"getCurrentUser","responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}},"description":"OK"}},"tags":["user-controller"]}},"/api/users/profile":{"get":{"operationId":"getProfile","responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}},"description":"OK"}},"tags":["user-controller"]},"put":{"operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileRequest"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseUserResponse"}}},"description":"OK"}},"tags":["user-controller"]}}},"components":{"schemas":{"ApiResponseAuthResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/AuthResponse"},"message":{"type":"string"}}},"ApiResponseJoinMeetingResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/JoinMeetingResponse"},"message":{"type":"string"}}},"ApiResponseListChatMessage":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ChatMessage"}},"message":{"type":"string"}}},"ApiResponseListMeeting":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Meeting"}},"message":{"type":"string"}}},"ApiResponseListMeetingRecording":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/MeetingRecording"}},"message":{"type":"string"}}},"ApiResponseListParticipantResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ParticipantResponse"}},"message":{"type":"string"}}},"ApiResponseMeeting":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/Meeting"},"message":{"type":"string"}}},"ApiResponseMeetingHistoryResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/MeetingHistoryResponse"},"message":{"type":"string"}}},"ApiResponseMeetingInfoResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/MeetingInfoResponse"},"message":{"type":"string"}}},"ApiResponseMeetingRecording":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/MeetingRecording"},"message":{"type":"string"}}},"ApiResponseMeetingSummaryResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/MeetingSummaryResponse"},"message":{"type":"string"}}},"ApiResponsePageMeetingHistoryResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/PageMeetingHistoryResponse"},"message":{"type":"string"}}},"ApiResponseString":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"string"},"message":{"type":"string"}}},"ApiResponseUserResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/UserResponse"},"message":{"type":"string"}}},"ApiResponseVoid":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{},"message":{"type":"string"}}},"ApprovalRequest":{"type":"object","properties":{"action":{"type":"string"},"participantId":{"type":"string","format":"uuid"}}},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserResponse"}}},"ChatMessage":{"type":"object","properties":{"content":{"type":"string"},"id":{"type":"string"},"meetingCode":{"type":"string"},"senderId":{"type":"string","format":"uuid"},"senderName":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"CreateMeetingRequest":{"type":"object","properties":{"access_type":{"type":"string","enum":["OPEN","TRUSTED","RESTRICTED"]},"end_time":{"type":"string","format":"date-time"},"participant_emails":{"type":"array","items":{"type":"string"}},"password":{"type":"string"},"settings":{"type":"string"},"start_time":{"type":"string","format":"date-time"},"title":{"type":"string"}}},"FeedbackRequest":{"type":"object","properties":{"rating":{"type":"integer","format":"int32"}}},"ForgotPasswordRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1,"pattern":"^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$"}},"required":["email"]},"GoogleLoginRequest":{"type":"object","properties":{"idToken":{"type":"string","minLength":1}},"required":["idToken"]},"JoinMeetingRequest":{"type":"object","properties":{"displayName":{"type":"string"},"password":{"type":"string"}}},"JoinMeetingResponse":{"type":"object","properties":{"currentHostId":{"type":"string"},"isOwner":{"type":"boolean"},"message":{"type":"string"},"owner":{"type":"boolean"},"role":{"type":"string"},"serverUrl":{"type":"string"},"settings":{"type":"string"},"status":{"type":"string"},"token":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1,"pattern":"^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$"},"password":{"type":"string","minLength":1}},"required":["email","password"]},"Meeting":{"type":"object","properties":{"accessType":{"type":"string","enum":["OPEN","TRUSTED","RESTRICTED"]},"allowedDomain":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"hostId":{"type":"string","format":"uuid"},"instant":{"type":"boolean"},"meetingCode":{"type":"string"},"meetingId":{"type":"string","format":"uuid"},"ownerId":{"type":"string","format":"uuid"},"password":{"type":"string"},"settings":{"type":"string"},"startTime":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["SCHEDULED","ACTIVE","FINISHED","CANCELED"]},"title":{"type":"string"}}},"MeetingHistoryResponse":{"type":"object","properties":{"canViewChatHistory":{"type":"boolean"},"canViewRecordings":{"type":"boolean"},"endTime":{"type":"string","format":"date-time"},"host":{"type":"boolean"},"meetingCode":{"type":"string"},"owner":{"type":"boolean"},"startTime":{"type":"string","format":"date-time"},"status":{"type":"string"},"title":{"type":"string"}}},"MeetingInfoResponse":{"type":"object","properties":{"access_type":{"type":"string","enum":["OPEN","TRUSTED","RESTRICTED"]},"host_name":{"type":"string"},"is_password_protected":{"type":"boolean"},"meeting_code":{"type":"string"},"passwordProtected":{"type":"boolean"},"status":{"type":"string"},"title":{"type":"string"}}},"MeetingRecording":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"egressId":{"type":"string"},"fileUrl":{"type":"string"},"id":{"type":"integer","format":"int64"},"meetingId":{"type":"string","format":"uuid"},"ownerId":{"type":"string","format":"uuid"},"roomName":{"type":"string"},"status":{"type":"string"}}},"MeetingSummaryResponse":{"type":"object","properties":{"duration":{"type":"string"},"messages":{"type":"integer","format":"int32"},"participants":{"type":"integer","format":"int32"}}},"PageMeetingHistoryResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/MeetingHistoryResponse"}},"empty":{"type":"boolean"},"first":{"type":"boolean"},"last":{"type":"boolean"},"number":{"type":"integer","format":"int32"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"size":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"sort":{"$ref":"#/components/schemas/SortObject"},"unpaged":{"type":"boolean"}}},"ParticipantResponse":{"type":"object","properties":{"avatarUrl":{"type":"string"},"displayName":{"type":"string"},"email":{"type":"string"},"participantId":{"type":"string","format":"uuid"},"requestTime":{"type":"string"},"status":{"type":"string"},"userId":{"type":"string","format":"uuid"}}},"RegisterRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1,"pattern":"^[A-Za-z0-9+_.-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$"},"fullName":{"type":"string","maxLength":100,"minLength":2},"password":{"type":"string","maxLength":2147483647,"minLength":8}},"required":["email","fullName","password"]},"ResetPasswordRequest":{"type":"object","properties":{"newPassword":{"type":"string","maxLength":2147483647,"minLength":8},"token":{"type":"string","minLength":1}},"required":["newPassword","token"]},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}},"UpdateProfileRequest":{"type":"object","properties":{"avatarUrl":{"type":"string"},"fullName":{"type":"string","maxLength":100,"minLength":2}}},"UserResponse":{"type":"object","properties":{"authProvider":{"type":"string"},"avatarUrl":{"type":"string"},"email":{"type":"string"},"fullName":{"type":"string"},"userId":{"type":"string","format":"uuid"}}}}}}